Many improvements for Spotify Web API
* Simplified authorization method: the auth token isn't saved anymore.
Instead, it's generated with the refresh token always.
* Modularized the credentials prompt: moved most of the Spotify Web
methods in MainWindow to a separate module.
* Many other small fixes
